pyproject.toml
[tool.poetry]
name = "alyautdinov-ruslan-hw-38"
version = "0.1.0"
description = ""
authors = ["Аляутдинов Руслан <rusalyautdinov@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2 = "^2.9.10"
pillow = "^11.0.0"
python-dotenv = "^1.0.1"
django = "^5.1.4"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.4.0"


[tool.poetry.group.lint.dependencies]
mypy = "^1.14.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
line-length = 119

[tool.isort]
line_length = 119